feat: add excluded_checks option
This feature allows users to pass a Tuple of check functions that
they want to be excluded. Exclusions are applied as a final step
so adding an additional_check and excluding it will result in the
check not being run.

This feature can also be utilised in the cli by passing a comma
separated list of the checks they want to exclude to the
`--exclude-checks` argument