fix(oas3.1): correctly validate paths using $ref
Schemathesis uses static version of the current published version
of the OpenApi 3.1 spec metaschema
(https://spec.openapis.org/oas/3.1/schema/2022-10-07)
to validate openapi 3.1 spec documents. Unfortunately, the published
version has at least one known bug in which the schema for `paths`
references the definition of a concrete `path-item` instead of
`path-item-or-reference`, which might still be technically incorrect
when it comes handling the case of ref and sibling fields, but is
correct according to the documented definition of a pathItemObject.

This oversight has been noticed multiple times
https://github.com/OAI/OpenAPI-Specification/issues/3298
https://github.com/OAI/OpenAPI-Specification/issues/2635#issuecomment-960248801
https://github.com/OAI/OpenAPI-Specification/issues/2635#issuecomment-872893673
https://github.com/OAI/OpenAPI-Specification/issues/3513
https://github.com/OAI/OpenAPI-Specification/pull/2657#issuecomment-2016588850

And finally fixed in Feb 2024
https://github.com/OAI/OpenAPI-Specification/pull/3355
with a slightly bigger rework of the pathItem schema.

Sadly, due to confusion about how to release fixes in schemas
https://github.com/OAI/OpenAPI-Specification/issues/151#issuecomment-1556268486
this change has not been published anywhere except schema.yaml in the
git repo, not even in schema.json, which appearantly only gets refreshed
once per release of the metaschema
https://github.com/OAI/OpenAPI-Specification/pull/3355#issuecomment-1690686993

This commit updates the stored schema from the most up-to-date 3.1.0
schema.yaml from 0035208 to close the bug and make spec-valid openapi
spec files that use $ref under path finally validate correctly in
schemathesis. It also adds a corresponding regression test